# server/myrti_core/src/model/asset_base.rs
use std::{fmt::Display, str::FromStr};

use camino::Utf8PathBuf as PathBuf;
use chrono::{DateTime, FixedOffset, Utc};
use eyre::eyre;
use serde::Serialize;

use crate::model::{FileId, MirrorCorrection, RotationCorrection};

use super::{AssetId, AssetRootDirId, AssetType};

#[derive(Debug, Clone, PartialEq, Eq, Hash)]
pub struct AssetFile {
    pub id: FileId,
    pub ty: AssetType,
    pub root_dir_id: AssetRootDirId,
    pub file_type: String,
    pub file_path: PathBuf,
    pub added_at: DateTime<Utc>,
    pub size: Size,
    pub rotation_correction: RotationCorrection,
    pub mirror_correction: MirrorCorrection,
    /// Seahash of the file, if already computed
    pub hash: Option<u64>,
}

#[derive(Debug, Clone, PartialEq, Eq, Hash)]
pub struct AssetBase {
    pub id: AssetId,
    pub rep_file_id: FileId,
    pub ty: AssetType,
    pub is_hidden: bool,
    pub taken_date: DateTime<Utc>,
    pub timestamp_info: TimestampInfo,
    pub gps_coordinates: Option<GpsCoordinates>,
}

/// Origin and reliability of the timezone for an asset's original creation date
#[derive(Debug, Clone, PartialEq, Eq, Hash)]
pub enum TimestampInfo {
    /// Asset metadata contained timezone information
    TzCertain(FixedOffset),
    /// Asset metadata contained UTC timestamp with no timezone information
    UtcCertain,
    /// Timezone was specified by the user
    TzSetByUser(FixedOffset),
    /// Timezone was inferred from location tags
    TzInferredLocation(FixedOffset),
    /// Asset metadata contained local date and time with no timezone info,
    /// so it was assumed to be the local timezone of this system
    TzGuessedLocal(FixedOffset),
    /// No timestamp in asset metadata, the timestamp is made up
    NoTimestamp,
}

#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Hash)]
pub struct Size {
    pub width: i32,
    pub height: i32,
}

#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
pub enum ThumbnailType {
    SmallSquare,
    LargeOrigAspect,
}

#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
pub enum ThumbnailFormat {
    Webp,
    Avif,
}

#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
pub struct GpsCoordinates {
    /// multiplied by 10e8
    pub lat: i64,
    /// multiplied by 10e8
    pub lon: i64,
}

impl AssetBase {
    pub fn taken_date_local(&self) -> DateTime<FixedOffset> {
        match self.timestamp_info {
            TimestampInfo::NoTimestamp => self.taken_date.into(),
            TimestampInfo::UtcCertain => self.taken_date.into(),
            TimestampInfo::TzCertain(tz)
            | TimestampInfo::TzSetByUser(tz)
            | TimestampInfo::TzInferredLocation(tz)
            | TimestampInfo::TzGuessedLocal(tz) => self.taken_date.with_timezone(&tz),
        }
    }
}

impl FromStr for ThumbnailFormat {
    type Err = eyre::Report;

    fn from_str(s: &str) -> Result<Self, Self::Err> {
        match s {
            "webp" => Ok(ThumbnailFormat::Webp),
            "avif" => Ok(ThumbnailFormat::Avif),
            other => Err(eyre!("Can't parse unkown thumbnail format: {}", other)),
        }
    }
}

impl Display for ThumbnailFormat {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        write!(
            f,
            "{}",
            match self {
                ThumbnailFormat::Avif => "avif",
                ThumbnailFormat::Webp => "webp",
            }
        )
    }
}