# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "myrti",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/albums": {
      "get": {
        "tags": [
          "crate::routes::album"
        ],
        "operationId": "getAllAlbums",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Album"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::album"
        ],
        "operationId": "createAlbum",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlbumRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAlbumResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assets/timeline": {
      "get": {
        "tags": [
          "crate::routes::timeline"
        ],
        "operationId": "getTimeline",
        "parameters": [
          {
            "name": "lastAssetId",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/AssetId"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "maxCount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "lastFetch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineChunk"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/photoSeries": {
      "post": {
        "tags": [
          "crate::routes::photo_series"
        ],
        "operationId": "createSeries",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSeriesResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/timelinegroups": {
      "post": {
        "tags": [
          "crate::routes::timeline_group"
        ],
        "operationId": "createTimelineGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTimelineGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTimelineGroupResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "crate::routes::timeline_group"
        ],
        "operationId": "editTimelineGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTimelineGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Album": {
        "type": "object",
        "required": [
          "id",
          "numAssets",
          "createdAt",
          "changedAt"
        ],
        "properties": {
          "changedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "$ref": "#/components/schemas/AlbumId"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "numAssets": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "AlbumId": {
        "type": "string"
      },
...
      "Asset": {
        "type": "object",
        "required": [
          "assetId",
          "takenDate",
          "repFile"
        ],
        "properties": {
          "assetId": {
            "$ref": "#/components/schemas/AssetId"
          },
          "repFile": {
            "$ref": "#/components/schemas/AssetFile"
          },
          "takenDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AssetFile": {
        "type": "object",
        "required": [
          "fileId",
          "assetRootId",
          "pathInRoot",
          "width",
          "height",
          "addedAt",
          "rotationCorrection",
          "mirrorCorrection",
          "mimeType"
        ],
        "properties": {
          "addedAt": {
            "type": "string",
            "format": "date-time"
          },
          "assetRootId": {
            "$ref": "#/components/schemas/AssetRootDirId"
          },
          "fileId": {
            "$ref": "#/components/schemas/FileId"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "mimeType": {
            "type": "string"
          },
          "mirrorCorrection": {
            "$ref": "#/components/schemas/MirrorCorrection"
          },
          "pathInRoot": {
            "type": "string"
          },
          "rotationCorrection": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AssetId": {
        "type": "string"
      },
      "AssetRootDirId": {
        "type": "string"
      },
...
      "AssetSeriesId": {
        "type": "string"
      },
      "AssetSpe": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Image"
              },
              {
                "type": "object",
                "required": [
                  "assetType"
                ],
                "properties": {
                  "assetType": {
                    "type": "string",
                    "enum": [
                      "image"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Video"
              },
              {
                "type": "object",
                "required": [
                  "assetType"
                ],
                "properties": {
                  "assetType": {
                    "type": "string",
                    "enum": [
                      "video"
                    ]
                  }
                }
              }
            ]
          }
        ],
        "discriminator": {
          "propertyName": "assetType"
        }
      },
...
      "AssetWithSpe": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Asset"
          },
          {
            "$ref": "#/components/schemas/AssetSpe"
          },
          {
            "type": "object"
          }
        ]
      },
      "CreateAlbumRequest": {
        "type": "object",
        "required": [
          "name",
          "assets"
        ],
        "properties": {
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetId"
            }
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CreateAlbumResponse": {
        "type": "object",
        "required": [
          "albumId"
        ],
        "properties": {
          "albumId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CreateSeriesRequest": {
        "type": "object",
        "required": [
          "assetIds"
        ],
        "properties": {
          "assetIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetId"
            }
          }
        }
      },
      "CreateSeriesResponse": {
        "type": "object",
        "required": [
          "seriesId"
        ],
        "properties": {
          "seriesId": {
            "$ref": "#/components/schemas/AssetSeriesId"
          }
        }
      },
      "CreateTimelineGroupRequest": {
        "type": "object",
        "required": [
          "assets",
          "name"
        ],
        "properties": {
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetId"
            }
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CreateTimelineGroupResponse": {
        "type": "object",
        "required": [
          "timelineGroupId",
          "displayDate"
        ],
        "properties": {
          "displayDate": {
            "type": "string",
            "format": "date-time"
          },
          "timelineGroupId": {
            "$ref": "#/components/schemas/TimelineGroupId"
          }
        }
      },
...
      "EditTimelineGroup": {
        "type": "string",
        "enum": [
          "add",
          "remove"
        ]
      },
      "EditTimelineGroupRequest": {
        "type": "object",
        "required": [
          "assets",
          "groupId",
          "operation"
        ],
        "properties": {
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetId"
            }
          },
          "groupId": {
            "$ref": "#/components/schemas/TimelineGroupId"
          },
          "operation": {
            "$ref": "#/components/schemas/EditTimelineGroup"
          }
        }
      },
      "FileId": {
        "type": "string"
      },
...
      "Image": {
        "type": "object",
        "required": [
          "representations"
        ],
        "properties": {
          "representations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageRepresentation"
            }
          }
        }
      },
      "ImageRepresentation": {
        "type": "object",
        "required": [
          "id",
          "format",
          "width",
          "height",
          "size"
        ],
        "properties": {
          "format": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "MirrorCorrection": {
        "type": "string",
        "enum": [
          "None",
          "Horizontal",
          "Vertical"
        ]
      },
...
      "TimelineChunk": {
        "type": "object",
        "description": "Response for a request for the next part of the timeline to display\n\n`groups` are always whole, not sliced in the middle. Either TimelineGroup or Day\n`date` is the date before queries are made",
        "required": [
          "date",
          "changedSinceLastFetch",
          "groups"
        ],
        "properties": {
          "changedSinceLastFetch": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineGroup"
            }
          }
        }
      },
      "TimelineGroup": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TimelineGroupType"
          },
          {
            "type": "object",
            "required": [
              "assets"
            ],
            "properties": {
              "assets": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AssetWithSpe"
                }
              }
            }
          }
        ]
      },
      "TimelineGroupId": {
        "type": "string"
      },
      "TimelineGroupType": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "date",
              "type"
            ],
            "properties": {
              "date": {
                "type": "string",
                "format": "date"
              },
              "type": {
                "type": "string",
                "enum": [
                  "day"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "groupTitle",
              "groupStartDate",
              "groupEndDate",
              "groupId",
              "type"
            ],
            "properties": {
              "groupEndDate": {
                "type": "string",
                "format": "date-time"
              },
              "groupId": {
                "type": "string"
              },
              "groupStartDate": {
                "type": "string",
                "format": "date-time"
              },
              "groupTitle": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "group"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "Video": {
        "type": "object",
        "required": [
          "hasDash"
        ],
        "properties": {
          "hasDash": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}